Use same array as Sample or TinySample depending on sampling rate
Instead of defining the dataBuffer as TinySample and converting
while sending back, this version keeps the array as is so that the
normal low sampling frequency run proceeds as before.

For high frequency runs, every timer count sample is collected
and saved in the same databuffer but cast to TinySample *.
TinySample now contains only volt and curr (4 bytes) allowing for
4 times as many sample (all buffered until the end of run).
Time in TinySample is redundant as it can be recovered from
the index. Channel index info is lost as its not saved in
tiny sample.

- Channel number in Sample is changed to 4 bit uint as there is no
memory cost. This also keeps Sample array to have no padding eventhough
the cast to TinySample may still work without doing so.

Signed-off-by: Selva Nair <selva.nair@gmail.com>