Add a speed test during startup
The result (# of updateTest calls per usec) is returned
when queried for firmware with json key "timeElapsed1" and
"timeElapsed2" for original updateTestTimer and new
updateTestTimerFast called 1000 times.

During tests, same variables are used to count teh actual time
evolved between start and end test. timeElapsed1 used when original
update routine is used (< 1000 Hz sampling) and timeElasped2
used when the new routine (faster sampling) is used. Again,
these number are included only in response to firmware, so
use that afte the run to check it. The numbers are in usec.

Signed-off-by: Selva Nair <selva.nair@gmail.com>